move tickets table
